+++ fractal_server/app/routes/api/v2/_aux_functions_task_lifecycle.py
@@ -4,7 +4,11 @@ from fastapi import HTTPException
 from fastapi import status
 from httpx import AsyncClient
 from httpx import TimeoutException
+from sqlmodel import select
 
+from fractal_server.app.db import AsyncSession
+from fractal_server.app.models.v2 import TaskGroupActivityV2
+from fractal_server.app.schemas.v2 import TaskGroupActivityStatusV2
 from fractal_server.logger import set_logger
 
 
@@ -122,3 +126,42 @@ async def get_package_version_from_pypi(
         # Case 3: `version` is unset and we use latest
         logger.info(f"No version requested, returning {latest_version=}.")
         return latest_version
+
+
+async def check_no_ongoing_activity(
+    *,
+    task_group_id: int,
+    db: AsyncSession,
+) -> None:
+    """
+    Find ongoing activities for the same task group.
+
+    Arguments:
+        task_group_id:
+        db:
+    """
+    # DB query
+    stm = (
+        select(TaskGroupActivityV2)
+        .where(TaskGroupActivityV2.taskgroupv2_id == task_group_id)
+        .where(TaskGroupActivityV2.status == TaskGroupActivityStatusV2.ONGOING)
+    )
+    res = await db.execute(stm)
+    ongoing_activities = res.scalars().all()
+
+    if ongoing_activities == []:
+        # All good, exit
+        return
+
+    msg = "Found ongoing activities for the same task-group:"
+    for ind, activity in enumerate(ongoing_activities):
+        msg = (
+            f"{msg}\n{ind + 1}) "
+            f"Action={activity.action}, "
+            f"status={activity.status}, "
+            f"timestamp_started={activity.timestamp_started}."
+        )
+    raise HTTPException(
+        status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
+        detail=msg,
+    )