+++ fractal_server/app/schemas/v2/task_group.py
@@ -85,19 +85,15 @@ class TaskGroupReadV2(BaseModel):
     pip_extras: Optional[str] = None
     pinned_package_versions: dict[str, str] = Field(default_factory=dict)
 
+    venv_size_in_kB: Optional[int] = None
+    venv_file_number: Optional[int] = None
+
     active: bool
     timestamp_created: datetime
 
 
 class TaskGroupUpdateV2(BaseModel, extra=Extra.forbid):
     user_group_id: Optional[int] = None
-    active: Optional[bool] = None
-
-    @validator("active")
-    def active_cannot_be_None(cls, value):
-        if value is None:
-            raise ValueError("`active` cannot be set to None")
-        return value
 
 
 class TaskGroupActivityV2Read(BaseModel):