+++ fractal_server/tasks/v2/ssh/deactivate.py
@@ -1,0 +1,2 @@
+def deactivate_ssh():
+    pass