+++ fractal_server/tasks/v2/ssh/reactivate.py
@@ -1,0 +1,2 @@
+def reactivate_ssh():
+    pass